Keep the c1-ffl circuit config free of React
utils.ts imports CIRCUIT_CONFIG and was transitively pulling the Flow
node components with it; nodeTypes moves to nodes.ts so the constants
stay data. The lodash subpath imports are explicit rather than smaller —
measured byte-identical either way.
