+++ pages/blocks/c1-ffl/figures/nodes.ts
@@ -1,0 +1,15 @@
+import {
+  CircuitProteinNode,
+  LineNode,
+  CircleNode,
+  CircuitPromoterNode,
+  LabelNode,
+} from "@viz/flow";
+
+export const nodeTypes = {
+  circle: CircleNode,
+  protein: CircuitProteinNode,
+  promoter: CircuitPromoterNode,
+  line: LineNode,
+  label: LabelNode,
+};
